requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
python-dotenv
google-generativeai
PyMySQL
psycopg2-binary
